Restricted Pages

+++ studybud/base/views.py
@@ -1,7 +1,9 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
@@ -19,9 +21,8 @@ def logoutUser(request):
     return redirect("home")
 
 def loginPage(request):
-    username = ""
-    password = ""
-    user = None
+    if request.user.is_authenticated:
+        return redirect("home")
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -40,7 +41,7 @@ def loginPage(request):
             messages.error(request, "Username OR password does not exist")
 
 
-    context = {"user": user}
+    context = {}
     return render(request, 'base/login_register.html', context)
 
 def home(request):
@@ -62,6 +63,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url="/login")
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -73,11 +75,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url="/login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid:
@@ -87,8 +92,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="/login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
